♻️ Move config and particle generation into separate files
